api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9
elasticsearch==8.12.1
httpx==0.26.0
openai>=1.3.0
socksio>=1.0.0
python-dotenv==1.0.1
tiktoken==0.6.0
jieba==0.42.1
pypdf==3.17.4
python-docx==1.1.0
